ONS NA 2019 demo web server
This is the webserver for the ONS NA 2019 demo. The demo writes license
plate numbers to an Influx database, and this simple web server can be used
to browse the structure of the database and to display the contents of a
chosen table, with time constraints (eg. show last 1 minute of data).

Change-Id: If26cb886c5d83c46c4bb879f84f73905ba257c84
Signed-off-by: Tapio Tallgren <tapio.tallgren@nokia.com>