WIP: uMEC in a container
This is the final patch in this series, since I am abandoning this
approach. The goal was to run the uMEC stack in a container to be able
to test it automatically. This approach has three stages:

1. The builder has Rust installed and can compile Rust programs in
Tumbleweed, using the ONBUILD magic

2. The second stage starts with plain Tumbleweed and installs systemd to
it.

3. The third stage installs k3s

It is possible to run containerd with podman inside the container,
systemd works, and k3s installs just fine. Difficulty is that k3s and
k8s expect to use different kernel modules and have the route_localnet
option on. This means that either the container must run in privileged
mode (which is scary with systemd in the container), or that the host
must be modified as root. An easier approach is to launch a virtual
machine.

Anyway, I am storing this work as it may turn useful later.

Change-Id: I8b3559c227f1e0e5883dccf84af9d04f9adfddac
Signed-off-by: Tapio Tallgren <tapio.tallgren@nokia.com>